# models/todo.go
package models

import (
	"errors"
	"strings"
	"time"
)

// CreateTodoRequest represents the input payload for creating a todo
type CreateTodoRequest struct {
	Title string `json:"title" example:"Learn Clean Architecture in Go"`
}

type Todo struct {
	ID        int       `json:"id"`
	UserID    int       `json:"user_id"`
	Title     string    `json:"title"`
	Completed bool      `json:"completed"`
	CreatedAt time.Time `json:"created_at"`
}

func (t *Todo) Validate() error {
	t.Title = strings.TrimSpace(t.Title)
	if t.Title == "" {
		return errors.New("title is required")
	}
	if len(t.Title) < 3 {
		return errors.New("title must be at least 3 characters")
	}
	return nil
}